Fix: rename directory, added docs uses cases
