+++ api/modules/generate-excel/ts/index.ts
@@ -1,5 +1,5 @@
 import { Request, Response, Application, Router } from "express";
-import { ExcelHandler } from "@bg/excel/handler";
+import { Excel } from "@bg/excel/excel";
 import { IReturnHandler } from "./types";
 export /*bundle*/
   class Controller {
@@ -16,7 +16,7 @@ export /*bundle*/
     res: Response
   ): Promise<Response<IReturnHandler, Record<string, IReturnHandler>>> => {
     try {
-      const excelHandler: ExcelHandler = new ExcelHandler();
+      const excel: Excel = new Excel();
       const params = req.body;
 
       if (!params?.sheetData) throw "invalid sheetData, this is required";
@@ -40,7 +40,7 @@ export /*bundle*/
       };
 
       // Crea el archivo Excel
-      const result: IReturnHandler = await excelHandler.createExcel(
+      const result: IReturnHandler = await excel.create(
         specs
       );
       if (!result.status && Array.isArray(result.error))
