+++ library/babel.config.js
