+++ library/modules/excel/module.json
@@ -1,5 +1,5 @@
 {
-    "name": "handler",
+    "name": "excel",
     "ts": {
         "path": "ts",
         "files": [
