+++ library/modules/excel/test/read-excel.test.ts
@@ -1,37 +1,36 @@
-import { ExcelHandler } from "../ts/index";
-import * as Excel from "exceljs";
+import { Excel } from "../ts/index";
 
-describe("ExcelHandler", () => {
-    let excelHandler;
+describe("Excel", () => {
+    let excel: Excel;
 
     beforeEach(() => {
-        excelHandler = new ExcelHandler();
+        excel = new Excel();
     });
 
-    describe('readExcel', () => {
+    describe('read', () => {
 
         it('debería leer un archivo Excel xlsx correctamente con readExcel', async () => {
             const filePath = './assets/test.xlsx';
             const params: any = { filePath, type: 'xlsx' };
 
-            const result = await excelHandler.readExcel(params);
+            const result = await excel.read(params);
 
             expect(result.status).toBe(true);
 
         });
 
-        it('debería lanzar un error si el tipo de archivo es incorrecto en readExcel', async () => {
+        it('debería lanzar un error si el tipo de archivo es incorrecto en read', async () => {
             const filePath = './assets/test.csv';
             const params: any = { filePath, type: 'invalid' };
 
-            await expect(excelHandler.readExcel(params)).rejects.toThrow('Type must be xlsx or csv');
+            await expect(excel.read(params)).rejects.toThrow('Type must be xlsx or csv');
         });
 
 
 
         it('debería leer un archivo Excel xlsx válido con validación de celdas y columnas', async () => {
             const filePath = './assets/test.xlsx'; // Ruta al archivo Excel xlsx válido
-            const params = {
+            const params: any = {
                 filePath,
                 type: 'xlsx',
                 cellsValidations: {
@@ -49,14 +48,14 @@ describe("ExcelHandler", () => {
                     ],
                 },
             };
-            const result = await excelHandler.readExcel(params);
+            const result = await excel.read(params);
             expect(result.status).toBe(true);
-            // Agregar más comprobaciones según sea necesario
+
         });
 
         it('debería manejar un error al leer un excel con celdas y columnas no válidas', async () => {
             const filePath = './assets/test.xlsx'; // Ruta al archivo Excel xlsx inválido
-            const params = {
+            const params: any = {
                 filePath,
                 type: 'xlsx',
                 cellsValidations: {
@@ -74,12 +73,10 @@ describe("ExcelHandler", () => {
                     ],
                 },
             };
-            const result = await excelHandler.readExcel(params);
+            const result = await excel.read(params);
             expect(result.status).toBe(false);
-            // Agregar más comprobaciones según sea necesario
-        });
 
-        // Agregar más pruebas para readExcel
+        });
     });
