+++ library/modules/excel/ts/validate-cells/utils/error-msgs.ts
@@ -1,0 +1,7 @@
+// Objeto con mensajes de error correspondientes a cada tipo
+export const errorMsgs: { [key: string]: string } = {
+    string: 'should be a string',
+    number: 'should be a number',
+    boolean: 'should be a boolean',
+    date: 'should be a Date object',
+};
