+++ readme.md
@@ -121,84 +121,7 @@ Estas propiedades permiten especificar las validaciones a aplicar a las celdas d
 
 ## Uso
 
-```javascript
-import { ExcelHandler } from "@bg/excel/handler";
-
-// Crear una instancia de ExcelHandler
-const excelHandler = new ExcelHandler();
-
-// Crear un archivo Excel (XLSX o CSV)
-const params = {
-    pathname: "output/",
-    options: {
-        // Opciones específicas del tipo de archivo
-        // ...
-    },
-    filename: "example.xlsx",
-    sheetData: [
-        {
-            sheetName: "Hoja1",
-            data: [
-                { name: "Juan", age: 25 },
-                { name: "María", age: 30 },
-            ],
-            columnsHeader: [
-                // Opcional
-                { header: "Nombre", key: "name" },
-                { header: "Edad", key: "age" },
-            ],
-        },
-        // Agregar más hojas si es necesario
-    ],
-    type: "xlsx", // Tipo de archivo ('xlsx' o 'csv')
-};
-
-excelHandler.createExcel(params).then((result) => {
-    if (result.status) {
-        console.log(`Archivo creado con éxito en: ${result.data.pathFile}`);
-    } else {
-        console.error(`Error al crear el archivo: ${result.error}`);
-    }
-});
-
-// Leer datos desde un archivo Excel o CSV
-const readParams = {
-    filePath: "input/example.xlsx",
-    cellsValidations: {
-        // Opcional
-        columnValidations: [
-            {
-                sheetName: "Hoja1",
-                columnKey: "age",
-                dataType: "number",
-                regexPattern: "/^[0-9]+$/",
-            },
-        ],
-        cellRangeValidations: [
-            // Opcional
-            {
-                sheetName: "Hoja1",
-                startRow: 2,
-                endRow: 3,
-                startCol: 2,
-                endCol: 2,
-                dataType: "number",
-                regexPattern: "/^[0-9]+$/",
-            },
-        ],
-    },
-    type: "xlsx", // Tipo de archivo ('xlsx' o 'csv')
-};
-
-excelHandler.readExcel(readParams).then((result) => {
-    if (result.status) {
-        console.log("Datos leídos con éxito:");
-        console.log(result.data);
-    } else {
-        console.error(`Error al leer el archivo: ${result.error}`);
-    }
-});
-```
+Puede detallar en [Este archivo](./use-cases.md) los casos de uso
 
 ## Contribuciones
