+++ tests/template/application/_variables.scss
