+++ tests/template/application/custom-properties/_dark.scss
