+++ tests/template/application/custom-properties/_light.scss
