+++ tests/template/application/custom-properties/custom-properties.scss
