+++ tests/template/application/custom-properties/neutral-palette.scss
