+++ tests/template/application/custom-properties/palette.scss
