+++ tests/template/application/custom-properties/variables.scss
