+++ tests/template/application/styles.scss
